refactor(vision): move vision header to another component
